requirements.txt
# PCM 彩板表面缺陷检测系统依赖
# Core deep learning & detection
ultralytics==8.4.17       # YOLO11 (latest, includes yolo11s.pt)
torch>=2.0.0
torchvision>=0.15.0

# Image processing
opencv-python>=4.8.0
Pillow>=10.0.0
albumentations>=1.3.0

# Data & utilities
numpy>=1.24.0
PyYAML>=6.0
tqdm>=4.65.0

# Visualization & metrics
matplotlib>=3.7.0
seaborn>=0.12.0
scikit-learn>=1.3.0
